v.04

+++ CurrencyConverter/Currency.cs
@@ -14,10 +14,10 @@ namespace CurrencyConverter
         public string CharCode { get; set; }
         public int Nominal { get; set; }
         public string Name { get; set; }
-        public float Value { get; set; }
-        public float Previous { get; set; }
+        public double Value { get; set; }
+        public double Previous { get; set; }
 
-        public Currency(string charcode, string name, float value, int nominal)
+        public Currency(string charcode, string name, double value, int nominal)
         {
             CharCode = charcode;
             Name = name;
